requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.37
python-dotenv==1.0.1
psycopg2-binary~=2.9.3